PIM-5033: Replace abstract number localizer and presenter by concrete class
