+++ src/Pim/Component/Localization/Localizer/MetricLocalizer.php
@@ -9,7 +9,7 @@ namespace Pim\Component\Localization\Localizer;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MetricLocalizer extends AbstractNumberLocalizer
+class MetricLocalizer extends NumberLocalizer
 {
     /**
      * {@inheritdoc}
