+++ src/Pim/Component/Localization/Localizer/PriceLocalizer.php
@@ -11,7 +11,7 @@ use Symfony\Component\Validator\ConstraintViolationList;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class PriceLocalizer extends AbstractNumberLocalizer
+class PriceLocalizer extends NumberLocalizer
 {
     /**
      * {@inheritdoc}
