+++ src/Pim/Component/Localization/Presenter/MetricPresenter.php
@@ -2,8 +2,6 @@
 
 namespace Pim\Component\Localization\Presenter;
 
-use Pim\Component\Localization\Factory\NumberFactory;
-
 /**
  * Metric presenter, able to render metric data readable for a human
  *
@@ -11,43 +9,19 @@ use Pim\Component\Localization\Factory\NumberFactory;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MetricPresenter implements PresenterInterface
+class MetricPresenter extends NumberPresenter
 {
-    /** @var NumberFactory */
-    protected $numberFactory;
-
-    /** @var string[] */
-    protected $attributeTypes;
-
-    /**
-     * @param NumberFactory $numberFactory
-     * @param string[]      $attributeTypes
-     */
-    public function __construct(NumberFactory $numberFactory, array $attributeTypes)
-    {
-        $this->numberFactory  = $numberFactory;
-        $this->attributeTypes = $attributeTypes;
-    }
-
     /**
      * {@inheritdoc}
      */
     public function present($value, array $options = [])
     {
-        $numberFormatter = $this->numberFactory->create($options);
-        if (isset($options['disable_grouping_separator']) && true === $options['disable_grouping_separator']) {
-            $numberFormatter->setSymbol(\NumberFormatter::GROUPING_SEPARATOR_SYMBOL, '');
+        if (is_array($value) && isset($value['data'])) {
+            $amount = parent::present($value['data'], $options);
+
+            return sprintf('%s %s', $amount, $value['unit']);
         }
-        $amount = $numberFormatter->format($value['data']);
 
-        return sprintf('%s %s', $amount, $value['unit']);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supports($attributeType)
-    {
-        return in_array($attributeType, $this->attributeTypes);
+        return parent::present($value, $options);
     }
 }
